update category previous text is showing now

+++ src/components/dashboard/DbCat.js
@@ -5,12 +5,15 @@ import {connect} from 'react-redux';
 import {compose} from 'redux';
 import {firestoreConnect} from 'react-redux-firebase';
 import { deleteCategory } from '../../store/actions/categoryActions';
+import moment from 'moment';
 
 
 class DbCat extends Component {
 
     deleteCategory = (docId) =>{
-        this.props.deleteCategory(docId);
+        if (window.confirm("This will delete all subcategories under it. Are You Sure To Delete This Category Permanantly.")) {
+            this.props.deleteCategory(docId);
+        } 
     }
     
 
@@ -18,7 +21,6 @@ class DbCat extends Component {
         const categories = this.props.categories;
 
         return (
-
             <ul className="collection with-header cat">
                 <li className="collection-header card">
                     <h4 style={{'textAlign': 'center'}}>Admin controll panel
@@ -32,12 +34,21 @@ class DbCat extends Component {
                     categories && categories.map(category=>{
                         return (
                             <li className="collection-item cat-item card" key={category.id}>
-                                <div > {category.name}
-                                    <span className="secondary-content"  title="Delete this category"  style={{cursor:'pointer'}} onClick={(e=>this.deleteCategory(category.id))}><i className="material-icons">delete</i></span>
-                                    <Link to={'/dashboard/category/update/'+category.id}  className="secondary-content" title="Edit this category" ><i className="material-icons">edit</i></Link>
-                                    <Link to={'/dashboard/subcat/create/' + category.id + '/' + category.name} className="secondary-content" title="Add a subcategory unde it"><i className="material-icons">add</i></Link>
+                                <div > {category.name} 
+                                    <span style={{'fontSize':'13px', 'fontWeight':'400'}}>
+                                        <i>  Updated: {moment( category.createdAt.toDate() ).calendar() + ' by ' + category.admin }  </i>
+                                    </span>
+                                    <span className="secondary-content"  title={ "Delete " + category.name + " and it's subcategory" }  style={{cursor:'pointer'}} onClick={(e=>this.deleteCategory(category.id))}>
+                                        <i className="material-icons">delete</i>
+                                    </span>
+                                    <Link to={'/dashboard/category/update/'+category.id}  className="secondary-content" title={'Edit ' + category.name} >
+                                        <i className="material-icons">edit</i>
+                                    </Link>
+                                    <Link to={'/dashboard/subcat/create/' + category.id + '/' + category.name} className="secondary-content" title={'Add a sub category under ' + category.name}>
+                                        <i className="material-icons">add</i>
+                                    </Link>
                                 </div>
-                                <DbSubCat subcats={category.subcats} />
+                                <DbSubCat subcats={category.subcats} id={category.id} />
                             </li>
                         )
                     })
@@ -48,6 +59,7 @@ class DbCat extends Component {
 }
 
 const mapStateToProps = (state) =>{
+    console.log('from dbcat', state)
     return {
         categories: state.firestore.ordered.categories
     }
